Moved GenericEventStreamer to src/plugins
